Working on BMS required data

+++ ITMProject/Core/Inc/therm_temperature_lookup.h
@@ -9,6 +9,6 @@
 #define INC_THERM_TEMPERATURE_LOOKUP_H_
 #include <stdint.h>
 
-uint16_t binary_search(uint16_t adc_target);
+int16_t binary_search(uint32_t adc_target);
 
 #endif /* INC_THERM_TEMPERATURE_LOOKUP_H_ */
